requirements.txt
openai
supabase
moviepy
yt-dlp
python-dotenv
undetected-chromedriver
ffmpeg-python
schedule